moving custom logig from test to multipoll_example
